+++ src/Symfony/Component/HttpFoundation/Request.php
@@ -120,6 +120,11 @@ class Request
     /**
      * @var array
      */
+    protected $encodings;
+
+    /**
+     * @var array
+     */
     protected $acceptableContentTypes;
 
     /**
@@ -1450,14 +1455,13 @@ class Request
      * Gets a list of encodings acceptable by the client browser.
      *
      * @return array List of encodings in preferable order
-     *
-     * @api
      */
     public function getEncodings()
     {
         if (null !== $this->encodings) {
             return $this->encodings;
         }
+
         return $this->encodings = array_keys(AcceptHeader::fromString($this->headers->get('Accept-Encoding'))->all());
     }